+++ config/features/deprecated_flags.go
@@ -62,6 +62,16 @@ var (
 		Usage:  deprecatedUsage,
 		Hidden: true,
 	}
+	deprecatedInteropGenesisTimeFlag = &cli.Uint64Flag{
+		Name:   "interop-genesis-time",
+		Usage:  deprecatedUsage,
+		Hidden: true,
+	}
+	deprecatedInteropNumValidatorsFlag = &cli.Uint64Flag{
+		Name:   "interop-num-validators",
+		Usage:  deprecatedUsage,
+		Hidden: true,
+	}
 )
 
 // Deprecated flags for both the beacon node and validator client.
@@ -77,6 +87,8 @@ var deprecatedFlags = []cli.Flag{
 	deprecatedVerboseSigVerification,
 	deprecatedEnableDebugRPCEndpoints,
 	deprecatedEnableExperimentalState,
+	deprecatedInteropGenesisTimeFlag,
+	deprecatedInteropNumValidatorsFlag,
 }
 
 // deprecatedBeaconFlags contains flags that are still used by other components